pyproject.toml
[project]
name = "lacrei-saude"
version = "0.1.0"
description = ""
authors = [
    {name = "Luis Miguel",email = "luismiguelsantossouza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
